Move LookupMixIn
Add disables for function-level imports
